Add output file option to DB upgrade; update pipeline action
The command for running the database upgrade has been updated to include an `--output-file` parameter, specifying the path to the output file as `.\dist\migration.sql`. This ensures the output is saved to the specified file.

The Bitbucket Pipelines configuration now includes a new allowed value `generate-upgrade-script` for the `action` parameter, enabling the pipeline to execute this new action and providing more flexibility in database operations.