Enhance Demographic mapping and foreign key constraints
Updated the `DemographicMap` class to configure the `Id` property with precision and added an index for `FullName` in the `ContactInfo` entity.

Modified the foreign key constraint name between `Demographic` and `EmploymentType` in `ProfitSharingDbContextModelSnapshot.cs`.

Created a new migration `20250307181438_dupNameSsnIndex.cs` to drop the old foreign key constraint, create a new index on the `FULL_NAME` column in the `DEMOGRAPHIC` table, and add the updated foreign key constraint. The `Down` method restores the previous state.