Update memberType logic and pagination take value
Modified the `memberType` property to conditionally set its value based on the length of `badgeNumber`. If `badgeNumber` exists and its length is less than or equal to 6, `memberType` is set to "all"; otherwise, it defaults to the value from `masterInquiryRequestParams?.memberType` or "beneficiaries". Increased the `take` property in the `pagination` object from 10 to 25.