Update package versions and enhance sorting functionality
Updated package versions in `Directory.Packages.props` from `9.1.0` to `9.2.0` for `Aspire.Hosting.AppHost`, `Aspire.Hosting.NodeJs`, and `Aspire.Oracle.EntityFrameworkCore`, and from `1.11.0-beta.1` to `1.11.0-beta.2` for `OpenTelemetry.Instrumentation.Quartz`.

In `ForfeitGrid.tsx`, modified the import statement to include `ISortParams`, added a new state variable `sortParams`, and updated the `onSearch` function to utilize it. Introduced a `sortEventHandler` to manage sorting changes and updated the `handleSortChanged` prop of `DSMGrid`.

In `ForfeitSearchFilter.tsx`, adjusted the `triggerSearch` function to include sorting parameters in the pagination object, specifying a sort by `badgeNumber` in descending order.