Add indexes to improve query performance and update version
This commit introduces new indexes to the `Demographic`, `PayProfit`, and `ProfitDetail` entity configurations to enhance query performance. Notably, an index on `TerminationDate` was added to the `Demographic` table, along with new indexes for `EnrollmentId` and `ProfitYear` in the `PayProfit` table. The `ProfitDetail` table also received several new indexes, including combinations of `ProfitYear` and `ProfitCodeId`, as well as `ProfitYear` and `MonthToDate`.

Additionally, the `ProductVersion` annotation in the `ProfitSharingDbContextModelSnapshot` has been updated from "9.0.3" to "9.0.4". A new migration class `addPAY426Indexes` has been introduced to manage the creation and deletion of these indexes, ensuring the database schema is properly updated.