Enhance auditing functionality and update packages
Updated the `Demoulas.Common.Data.Contexts` package to version `2.4.5-Alpha1` and added a new `MassTransit` package reference. Introduced a `Messaging` folder for better organization and added using directives for new interceptors in `Program.cs`.

Refactored the `AddDatabaseServices` method to support an `AuditSaveChangesInterceptor`, and created entity configurations for `AuditChange` and `AuditEvent`. Implemented the `AuditSaveChangesInterceptor` to handle auditing of entity changes.

Added a migration `20250424172050_addAudit.cs` to create the necessary audit tables and updated the `ProfitSharingDbContextModelSnapshot.cs` accordingly. Enhanced the overall structure of the codebase to improve maintainability and clarity.