Update PROFIT_DETAIL for service credit and dates
This commit updates the `PROFIT_DETAIL` table to set the
`YEARS_OF_SERVICE_CREDIT` based on conditions related to
`Contribution` and `COMMENT_TYPE_ID`. It also modifies
`MONTH_TO_DATE` to 1 when it is currently 20. A comment
has been added to reference task PS-1147 and clarify the
purpose of the `CREATED_UTC` update for improved sorting
of transactions.