Enhance Chip component styling in PSDrawer
Added a `sx` prop to the `Chip` component to apply custom styles based on the `statusName` of the `subPage` or `page` object. The `backgroundColor` and `color` properties are conditionally set for "In Progress" and "On Hold" statuses, with a font weight of 500. This improves the visual representation of the `Chip` component.