Refactor TerminationGrid to use StartAndEndDateRequest
Updated the TerminationGrid component to replace the TerminationRequest type with StartAndEndDateRequest. Introduced useFiscalCalendarYear hook and added an effect to reset the page number based on resetPageFlag. Modified grid data handling and updated the API query to accommodate the new request type. Removed the TerminationRequest interface as it is no longer needed.