Add NavigationRole support to NavigationFaker and mocks
A new method `GetAllNavigationRoles` was added to the
`NavigationFaker` class to generate a predefined list of
`NavigationRole` objects with `Id`, `Name`, and `IsReadOnly`
properties.

In `MockDataContextFactory`, the `GetAllNavigationRoles` method
was used to create a mocked `DbSet` of `NavigationRole` objects,
which was set up in both `_profitSharingDbContext` and
`_profitSharingReadOnlyDbContext` to support testing.